Remove excess Player-speed-while-in-Crowd logic
This code should have been removed as a part of 93e551d, but it was
accidentally left in.

This is what I get for making a bunch of changes at once, stashing
them, and splitting-up all that work into atomic commits... instead of
just making atomic commits in the first place. Indeed, this excess
logic was not present in my initial, pre-stashed changes.

(Note to self: This is my first time performing commit re-ordering.
I'm 4 commits into some unpushed changes, 5 now counting this one,
and I wanted this fix to be as early in the commit graph as it can
be.)