Changes made during the 2020-11-20 Game Development Group meeting
This project ended up being the subject for most of Friday's Game
Development Group meeting. During that meeting, I was taught more about
rigidbodies and colliders (along with many things about tilemaps during
the earlier subject that meet), making changes to the project myself as
the meeting went along. I am considering applying these changes to the
whole game, rather than the small subset it was done to during the
meet.

Here is a list of (hopefully all) the changes in no particular order.
"The scene" refers to Stage01Scene:
* Take the overrides made to the PlayerSprite in the scene and apply
  them to the PlayerSprite prefab. While this was done to save the
  PlayerSprite's components before messing with them, these overrides
  should have been in the PlayerSprite anyway.
* Replace the Rigidbody on the first LuggageCart in the scene with a
  Ridgidbody2D.
* Disable all colliders inherited from the LuggageCart prefab on the
  first LuggageCart.
* Add a PolygonCollider2D, CircleCollider2D, and CompositeCollider2D to
  the first LuggageCart.
* Alter the first LuggageCart's Polygon2DCollider's polygon to better
  fit the the luggage cart's shape.
* Replace the Rigidbody on the PlayerSprite in the scene with a
  Rigidbody2D.
* Alter some properties of the Rigidbody2D used by the PlayerSprite in
  the scene.
* Replace the BoxCollider on the PlayerSprite in the scene with a
  PolygonCollider2D.
* Alter the PlayerSprite's PolygonCollider2D's polygon to better fit the
  PlayerSprite's shape (the initial, perfectly-fitting shape was
  automatically generated).
* Change the Floor to use a Rigidbody2D.
* Convert Movement.cs to use 2D versions of collision and trigger
  callbacks (and subsequently the types of their parameters), and
  rigidbody.
* Assign velocity when moving the player instead of assigning a
  position.

Thanks to Zach Martychenko, who led in teaching me, and to "AJ", who got
me involved with the campus's Game Development Group in the first place
a few weeks ago.