Assign a 0 velocity to the PlayerSprite when teleporting
This does not appear to work consistently, if it has any affect at all.
At first, I thought the reason for this was because the Player is moved
by assigning positions, and not by assigning velocities; this is true,
but vertical movement of the PlayerSprite is carried out with
AddForce()'s and gravity from the physics engine, not from assigned
positions.

This is the second of two commits that are trying to prevent Space as
causing the PlayerSprite to jump while he is within a Stairwell's
BoxCollider Trigger.