Introduce new "Ceiling Lamp" obstacle
The CeilingLamp, like the Ice Machine, serves to temporarily prevent the
player from moving the PlayerSprite (player character). When the player
character jumps upwards into the inside of the CeilingLamp (see the
only enabled PolygonCollider2D) the player character is electrocuted,
being frozen for a brief time.  Upon unfreezing, a downward velocity is
imparted on the player character, slightly slower than the velocity he
had going into the lamp.

After the CeilingLamp is broken once, it does not electrocute the player
if he jumps into it again. I thought this serve to be a weird,
interesting quirk for the game, but this behavior may be changed to
always (or by chance, maybe?) electrocute the player instead.

This commit also introduces a sprite sheet for the PlayerSprite.