Show disabled Keycap when player character has no soda can(s)
This should help better communicate to the player that pressing "Q" is,
and can only be, used to activate the motivate power-up via a soda can
in the player character's posession.

The keycap already goes disabled when the player is in the motivated
state so as to say "you can't press it again, it won't have any effect",
so I considered this change to compliment and go along with that idea.