Introduce the new "[Rolled] Towel" power-up, add related stuff to HUD
The Rolled Towel makes the player character immune to getting stunned
by an Ice Machine for a brief period of time.

A part of my original idea was disallowing the simultaneous use of both
Motivation and a Rolled Towel in order to keep things balanced. However,
in implementing (or was it during testing?) this new power-up, I
remembered that the player character can only slip while he's motivated;
a Rolled Towel therefore would only have an effect in that circumstance.
So I've decided that— for now —both power-ups can be used simultaneously.