Add "Please Wait" keycap; make changes to {Keycap,Playercam}.cs
Because the PlayerSprite currently has no animations for any power-ups,
the keycap serves as the only explicit visual cue to the player when a
power-up is in use. When player character was using the last power-up he
had, the keycap would be in the same state— disabled —during and after
the power-up's use; whereas if he had more than one of any given
power-up, it would transition from disabled during use to enabled after.
I have introduced a new keycap appearance that is intended to convey a
state of "you will be able to use this key, but you must wait until this
power-up has worn off".