Delete unused prefabs that can be quickly recreated
Overall, the prefabs being deleted in this commit lack enough substance:
they are not used anywhere, are not tagged, and consist only of a sprite
renderer, all of which means that each of these can be recreated **in
seconds** should/once they actually be needed.

The second issue is that some of these prefabs are/were being used
instead of their functional, already existing counterparts.

SodaMachineV2 (no longer in the tree) and IceMachineV2 (which will soon
be removed in a branch that I've made, it'll be merged later) **fall
under this criteria as well**, and were used instead of their actually
functional counterparts ("Vending Machine" and "Ice Machine",
respectively) in Stage03. Both of these "V2" prefabs were used there
presumably because they had the updated graphics, and/or it was assumed
they were functional, when instead they were functionless. (n.b. that I
made the SodaMachineV2 prefab fully functional, but only after it was no
longer used anywhere, and it was right before I deleted it from the
tree; before all of that, again, that prefab fell under criteria
outlined above).

The pre-existing prefabs should have been modified instead, so that way
all Vending Machine's and Ice Machine's would all have consistent
behavior and appearance. Instead, we got 2 different versions of 2
prefabs, along with assumption (or something).

Last, some of these don't make sense as prefabs, such as the
Player{Fast,RunsIntoDoor} ones; their images will be used in Animations
and AnimationControllers, not used as prefabs by themselves.

Overall, please put more thought into *whether or not you should* be
making a new prefab for each new graphic that's introduced.