Revert *all* Rotation, Z-{Pos,Scale} overrides in *every* stage; Adjust Z-pos in prefabs; Adjust Z-pos of non-prefab objects in every stage
So, this evening I noticed that overrides to Transforms are so granular
that you can Apply or Revert overrides for each individual axis for
Position, Rotation, and Scale. How can you tell if any of these are
overridden in a prefab instance when looking at the Inspector? Their
text is bold and/or the affected axis(axes) label(s) is(are) bold. And
how do you revert it (or them, in the case of an entire Position,
Rotation, or Scale tuple)?  Right-click, either on the
"Position"/"Rotation"/"Scale" text for the entire respective tuple, or
on the label or field for just one particular value, and choose "Revert"
from the context menu. To my knowledge, there is no keyboard shortcut
(or option to assign one) to Revert an override on whatever field
currently has keyboard or mouse focus. Let that sink in: there are up to
9 possible things that can be overridden in a Transform, and I just
described how to Revert it(them) like a cave man.

Let's connect this back to the project:
1) Virtually all, if not all, prefab instances in every stage had an
   override on its Z-Position, where inheriting this from its prefab
   would be greatly preferable so it's just set in one place, rather
   than dozens.
2) Then, virtually all, if not all, prefab instances in every stage had
   an override on its entire Rotation, which, frustratingly, seemed to
   be (0,0,0) in every case (from what I remember)... again, it would be
   greatly preferable to inherit this from the prefab, and what's odd is
   that none of the prefabs are rotated at the root-of-the-prefab level.
   So why is there an override using the exact same values?  Anyway...
3) Last, and I think this was only for Floor, Wall, and Ceiling
   instances, there were Z-Scale overrides on some objects... once
   again, the same sentiment: just like the others, these, in hindsight,
   should have been set once in their prefab and made sure to be never
   overridden in stages.

So, for great justice, I was compelled to fix this. Only X- and
Y-positions and X- and Y-scales would be spared. All other Transform
overrides would be obliterated.

However, Unity doesn't allow you to revert overrides across multiple
GameObjects at once— even if the overrides use the same values for the
same fields on all selected instances of the same prefab —for some
reason(s) that's('re) beyond me. So what did I have to do? **Hundreds
of clicks** along with **hundreds of presses** of the "R" key. (I just
decided to bite that bullet rather than look for an Editor Script that
may or may not accelerate this task.) First Z-Positions. Then Z-Scales.
Then Rotations. Each one (or 3 at once in the case of Rotations)
reverted by hand. Like a cave man. Thankfully, neither carpal tunnel
ever complained.

But the results were, err, are, wonderful. I could now, and in fact did,
make every stage appear uniform along the Z-axis by just changing all
the prefabs, rather than every single instance in every single stage.
Again, now only handfuls of changes are needed, rather than dozens.

As for corporeal GameObjects in the stages that were not prefabs, well,
this didn't apply, but I did arrange their Z-Positions by hand in
coordination with the edits made to other prefabs.

Everything is looking uniform along the Z-axis. There may be some tweaks
I have to do (or not, I have the Editor closed right now), but I am
satisfied, and I am certain (no, really) that I haven't introduced any
regressions or mistakes

This took, like, two hours, a good chunk of it spent this wall of text.
But I am not going to lie, after looking at the shortstat numbers ("x
files changed, y insertions(+), z deletions(-)") for these changes, I
feel like I've just played and mastered one of those clicker games.

P.S. I know it looks like these Transforms are being introduced to
prefabs for the very first time. Unity's YAML works in mysterious
ways...

P.P.S. Full disclosure: I am aware of the existence of local
{position,rotation,scale}, but I don't know enough about them to
immediately answer the question "how do they differ from world
{position,rotation,scale}?"